refactor: use columnadjuster from common repo
